Folder Processing of MPRs
New functionality added to provide both single file and folder processing.

+++ 5_GUI_SF_CE_SC.py
@@ -1,25 +1,46 @@
 import tkinter as tk
-from tkinter import filedialog
-from tkinter import messagebox
+from tkinter import filedialog, messagebox
 import os
 from galvani import BioLogic
 import pandas as pd
 import matplotlib.pyplot as plt
 
-def select_mpr_file():
+
+def select_single_mpr_file():
     file_path = filedialog.askopenfilename(filetypes=[("MPR files", "*.mpr")])
-    mpr_file_entry.delete(0, tk.END)
-    mpr_file_entry.insert(0, file_path)
+    if file_path:
+        mpr_file_entry.delete(0, tk.END)
+        mpr_file_entry.insert(0, file_path)
+        process_single_button.config(state=tk.NORMAL)
+        folder_option.set(False)
 
-def select_output_dir():
+
+def select_folder():
+    folder_path = filedialog.askdirectory()
+    if folder_path:
+        folder_entry.delete(0, tk.END)
+        folder_entry.insert(0, folder_path)
+        process_folder_button.config(state=tk.NORMAL)
+        folder_option.set(True)
+
+
+def select_single_output_dir():
     output_dir = filedialog.askdirectory()
-    output_dir_entry.delete(0, tk.END)
-    output_dir_entry.insert(0, output_dir)
+    single_output_dir_entry.delete(0, tk.END)
+    single_output_dir_entry.insert(0, output_dir)
 
-def process_data():
+
+def select_folder_output_dir():
+    output_dir = filedialog.askdirectory()
+    folder_output_dir_entry.delete(0, tk.END)
+    folder_output_dir_entry.insert(0, output_dir)
+
+
+def process_single_file():
     mpr_file_path = mpr_file_entry.get()
     mass = float(entry_mass.get())
-    output_dir = output_dir_entry.get()
+    output_dir = single_output_dir_entry.get()
+
     if not os.path.isfile(mpr_file_path):
         messagebox.showerror("Error", "Please select a valid MPR file.")
         return
@@ -27,102 +48,71 @@ def process_data():
         messagebox.showerror("Error", "Please select a valid output directory.")
         return
 
-        # Extracting filename from the path
     file_name = os.path.basename(mpr_file_path)
+    output_file_name = f"EZ-Proc_{os.path.splitext(file_name)[0]}"
 
-    # Appending 'EZ-Proc' to the beginning of the filename
-    output_file_name = f"EZ-Proc_{file_name}"
-
-    # Removing extension from the filename
-    output_file_name = os.path.splitext(output_file_name)[0]
-
-    # Load the Biologic ECLAB MPR file
     mpr_file = BioLogic.MPRfile(mpr_file_path)
-
-    # Extract data from the MPR file and convert to Pandas DataFrame
     df = pd.DataFrame(mpr_file.data)
 
-    # Identify the number of rows where Ns = 0
     num_rows_with_Ns_0 = len(df[df['Ns'] == 0])
-
-    # Skip the rows with Ns = 0 in the second dataframe
     df_rounded = df.iloc[num_rows_with_Ns_0:]
-
-    # Round the values in the DataFrame to two decimal places
     df_rounded = df_rounded.round(2)
 
-    # Initialize the third DataFrame with required columns
     df_third = pd.DataFrame(columns=['Ns', 'Q charge/discharge/mA.h', 'half cycle'])
 
-    # Initialize variables
     max_charge = 0
     half_cycle_check = 0
     row_loc = 1
 
-    # Loop through each row of the second DataFrame
     for index, row in df_rounded.iterrows():
-        # Check if half cycle matches with half_cycle_check
         if row['half cycle'] == half_cycle_check:
-            # Replace (Q-Qo)/mA.h value in the current row of the third DataFrame
             df_third.loc[row_loc, 'Ns'] = row['Ns']
             df_third.loc[row_loc, 'Q charge/discharge/mA.h'] = row['Q charge/discharge/mA.h']
             df_third.loc[row_loc, 'half cycle'] = row['half cycle']
         else:
-            # Move to the next row in the third DataFrame
             row_loc = row_loc + 1
             df_third.loc[row_loc, 'Ns'] = row['Ns']
             df_third.loc[row_loc, 'Q charge/discharge/mA.h'] = row['Q charge/discharge/mA.h']
             df_third.loc[row_loc, 'half cycle'] = row['half cycle']
-            # Update half_cycle_check and max_charge
             half_cycle_check = row['half cycle']
             max_charge = row['Q charge/discharge/mA.h']
 
-    # Save the third DataFrame to CSV
-    #df_third.to_csv(os.path.join(output_dir, f'{output_file_name}_third_dataframe.csv'), index=False, sep='\t', encoding='utf-8')
-
-    # Initialize the fourth DataFrame with required columns
-    df_fourth = pd.DataFrame(columns=['Cycle Number', 'Capa Charge (mA.h)', 'Capa Discharge (mA.h)', 'Specific Capacitancce (mA.h/gMA)', 'Coulombic Efficiency'])
+    df_fourth = pd.DataFrame(
+        columns=['Cycle Number', 'Capa Charge (mA.h)', 'Capa Discharge (mA.h)',
+                 'Specific Capacitance (mA.h/gMA)', 'Coulombic Efficiency'])
 
     cycle_number = 1
-    print (mass)
-    # Loop through each row of the fourth DataFrame
+
     for index in range(1, len(df_third), 2):
-        # Calculate Cycle Number and Efficiency based on consecutive rows in the third DataFrame
         capa_charge = df_third.loc[index, 'Q charge/discharge/mA.h']
         capa_discharge = df_third.loc[index + 1, 'Q charge/discharge/mA.h'] * (-1)
-        specific_capacitance= capa_discharge / mass
+        specific_capacitance = capa_discharge / mass
         coulombic_efficiency = capa_discharge / capa_charge * 100
-        # Round Specific Capacitance and Coulombic Efficiency to two decimal places
-        specific_capacitance= round(specific_capacitance, 2)
+        specific_capacitance = round(specific_capacitance, 2)
         coulombic_efficiency = round(coulombic_efficiency, 2)
-        # Add the calculated values to the fourth DataFrame
+
         df_fourth.loc[cycle_number, 'Cycle Number'] = cycle_number
         df_fourth.loc[cycle_number, 'Capa Charge (mA.h)'] = capa_charge
         df_fourth.loc[cycle_number, 'Capa Discharge (mA.h)'] = capa_discharge
-        df_fourth.loc[cycle_number, 'Specific Capacitancce (mA.h/gMA)'] = specific_capacitance
+        df_fourth.loc[cycle_number, 'Specific Capacitance (mA.h/gMA)'] = specific_capacitance
         df_fourth.loc[cycle_number, 'Coulombic Efficiency'] = coulombic_efficiency
         cycle_number += 1
 
-    # Print the fourth DataFrame
-    print("Fourth DataFrame (df_fourth):")
-    print(df_fourth)
+    df_fourth.to_csv(os.path.join(output_dir, f'{output_file_name}_dataframe.csv'), index=False, sep='\t',
+                     encoding='utf-8')
 
-
-    # Save the fourth DataFrame to CSV
-    df_fourth.to_csv(os.path.join(output_dir, f'{output_file_name}_dataframe.csv'), index=False, sep='\t', encoding='utf-8')
-
-    # Plot Efficiency vs Cycle Number
     plt.figure(figsize=(10, 5))
 
     plt.subplot(1, 2, 1)
-    plt.plot(df_fourth['Cycle Number'], df_fourth['Coulombic Efficiency'])
+    plt.plot(df_fourth['Cycle Number'], df_fourth['Coulombic Efficiency'], 'o', color='green')
     plt.xlabel('Cycle Number')
     plt.ylabel('Coulombic Efficiency')
     plt.title('Coulombic Efficiency vs Cycle Number')
+    plt.ylim(0, 110)
     plt.grid(True)
 
     plt.subplot(1, 2, 2)
-    plt.plot(df_fourth['Cycle Number'], df_fourth['Specific Capacitancce (mA.h/gMA)'])
+    plt.plot(df_fourth['Cycle Number'], df_fourth['Specific Capacitance (mA.h/gMA)'], 'o', color='blue')
     plt.xlabel('Cycle Number')
     plt.ylabel('Specific Capacitance (mA.h/gMA)')
     plt.title('Specific Capacitance (mA.h/gMA) vs Cycle Number')
@@ -130,44 +120,194 @@ def process_data():
 
     plt.tight_layout()
 
-    # Save plot to file
     output_plot_path = os.path.join(output_dir, f"{output_file_name}_data_analysis_plot.png")
     plt.savefig(output_plot_path)
 
-    plt.show()
+    messagebox.showinfo("Success", "Single MPR file processed successfully.")
+    clear_cache()
+
+
+def process_folder():
+    folder_path = folder_entry.get()
+    mass = float(entry_mass_folder.get())
+    output_dir = folder_output_dir_entry.get()
+
+    if not os.path.isdir(folder_path):
+        messagebox.showerror("Error", "Please select a valid folder.")
+        return
+    if not os.path.isdir(output_dir):
+        messagebox.showerror("Error", "Please select a valid output directory.")
+        return
+
+    for file_name in os.listdir(folder_path):
+        if file_name.endswith('.mpr'):
+            mpr_file_path = os.path.join(folder_path, file_name)
+            output_file_name = f"EZ-Proc_{os.path.splitext(file_name)[0]}"
+
+            mpr_file = BioLogic.MPRfile(mpr_file_path)
+            df = pd.DataFrame(mpr_file.data)
+
+            num_rows_with_Ns_0 = len(df[df['Ns'] == 0])
+            df_rounded = df.iloc[num_rows_with_Ns_0:]
+            df_rounded = df_rounded.round(2)
+
+            df_third = pd.DataFrame(columns=['Ns', 'Q charge/discharge/mA.h', 'half cycle'])
+
+            max_charge = 0
+            half_cycle_check = 0
+            row_loc = 1
+
+            for index, row in df_rounded.iterrows():
+                if row['half cycle'] == half_cycle_check:
+                    df_third.loc[row_loc, 'Ns'] = row['Ns']
+                    df_third.loc[row_loc, 'Q charge/discharge/mA.h'] = row['Q charge/discharge/mA.h']
+                    df_third.loc[row_loc, 'half cycle'] = row['half cycle']
+                else:
+                    row_loc = row_loc + 1
+                    df_third.loc[row_loc, 'Ns'] = row['Ns']
+                    df_third.loc[row_loc, 'Q charge/discharge/mA.h'] = row['Q charge/discharge/mA.h']
+                    df_third.loc[row_loc, 'half cycle'] = row['half cycle']
+                    half_cycle_check = row['half cycle']
+                    max_charge = row['Q charge/discharge/mA.h']
+
+            df_fourth = pd.DataFrame(
+                columns=['Cycle Number', 'Capa Charge (mA.h)', 'Capa Discharge (mA.h)',
+                         'Specific Capacitance (mA.h/gMA)', 'Coulombic Efficiency'])
+
+            cycle_number = 1
+
+            for index in range(1, len(df_third), 2):
+                capa_charge = df_third.loc[index, 'Q charge/discharge/mA.h']
+                capa_discharge = df_third.loc[index + 1, 'Q charge/discharge/mA.h'] * (-1)
+                specific_capacitance = capa_discharge / mass
+                coulombic_efficiency = capa_discharge / capa_charge * 100
+                specific_capacitance = round(specific_capacitance, 2)
+                coulombic_efficiency = round(coulombic_efficiency, 2)
+
+                df_fourth.loc[cycle_number, 'Cycle Number'] = cycle_number
+                df_fourth.loc[cycle_number, 'Capa Charge (mA.h)'] = capa_charge
+                df_fourth.loc[cycle_number, 'Capa Discharge (mA.h)'] = capa_discharge
+                df_fourth.loc[cycle_number, 'Specific Capacitance (mA.h/gMA)'] = specific_capacitance
+                df_fourth.loc[cycle_number, 'Coulombic Efficiency'] = coulombic_efficiency
+                cycle_number += 1
+
+            df_fourth.to_csv(os.path.join(output_dir, f'{output_file_name}_dataframe.csv'), index=False, sep='\t',
+                             encoding='utf-8')
+
+            plt.figure(figsize=(10, 5))
+
+            plt.subplot(1, 2, 1)
+            plt.plot(df_fourth['Cycle Number'], df_fourth['Coulombic Efficiency'], 'o', color='green')
+            plt.xlabel('Cycle Number')
+            plt.ylabel('Coulombic Efficiency')
+            plt.title('Coulombic Efficiency vs Cycle Number')
+            plt.ylim(0, 110)
+            plt.grid(True)
+
+            plt.subplot(1, 2, 2)
+            plt.plot(df_fourth['Cycle Number'], df_fourth['Specific Capacitance (mA.h/gMA)'], 'o', color='blue')
+            plt.xlabel('Cycle Number')
+            plt.ylabel('Specific Capacitance (mA.h/gMA)')
+            plt.title('Specific Capacitance (mA.h/gMA) vs Cycle Number')
+            plt.grid(True)
+
+            plt.tight_layout()
+
+            output_plot_path = os.path.join(output_dir, f"{output_file_name}_data_analysis_plot.png")
+            plt.savefig(output_plot_path)
+
+    messagebox.showinfo("Success", "All MPR files in the folder processed successfully.")
+    clear_cache()
+
+
+def clear_cache():
+    mpr_file_entry.delete(0, tk.END)
+    single_output_dir_entry.delete(0, tk.END)
+    folder_entry.delete(0, tk.END)
+    entry_mass.delete(0, tk.END)
+    entry_mass_folder.delete(0, tk.END)
+    folder_output_dir_entry.delete(0, tk.END)
+    process_single_button.config(state=tk.DISABLED)
+    process_folder_button.config(state=tk.DISABLED)
+    folder_option.set(False)
+
+
+def on_closing():
+    if messagebox.askokcancel("Quit", "Do you want to quit?"):
+        root.destroy()
+
 
 # Create UI
 root = tk.Tk()
-root.title("AM Single MPR File Processor")
+root.title("AM MPR File Processor")
 
-# MPR file selection
-mpr_file_label = tk.Label(root, text="Select MPR File:")
+folder_option = tk.BooleanVar(root, value=False)
+
+single_frame = tk.LabelFrame(root, text="Single File")
+single_frame.grid(row=0, column=0, padx=10, pady=10, sticky="nsew")
+
+folder_frame = tk.LabelFrame(root, text="Folder")
+folder_frame.grid(row=1, column=0, padx=10, pady=10, sticky="nsew")
+
+single_button = tk.Radiobutton(root, text="Single File", variable=folder_option, value=False, command=lambda: select_single_mpr_file())
+single_button.grid(row=2, column=0, padx=5, pady=5, sticky="w")
+
+folder_button = tk.Radiobutton(root, text="Folder", variable=folder_option, value=True, command=lambda: select_folder())
+folder_button.grid(row=2, column=0, padx=5, pady=5, sticky="e")
+
+# Single File Section
+mpr_file_label = tk.Label(single_frame, text="Select MPR File:")
 mpr_file_label.grid(row=0, column=0, sticky="w")
-mpr_file_entry = tk.Entry(root, width=50)
+mpr_file_entry = tk.Entry(single_frame, width=50)
 mpr_file_entry.grid(row=0, column=1, padx=5, pady=5)
-mpr_file_button = tk.Button(root, text="Browse", command=select_mpr_file)
+mpr_file_button = tk.Button(single_frame, text="Browse", command=select_single_mpr_file)
 mpr_file_button.grid(row=0, column=2, padx=5, pady=5)
 
-# Mass Label and Entry
-label_mass = tk.Label(root, text="Mass of Active Material (g):")
-label_mass.grid(row=1, column=0, sticky="w")
-entry_mass = tk.Entry(root, width=20)
-entry_mass.grid(row=1, column=1, sticky="w", padx=5)
+mass_label = tk.Label(single_frame, text="Mass of Active Material (g):")
+mass_label.grid(row=1, column=0, sticky="w")
+entry_mass = tk.Entry(single_frame, width=20)
+entry_mass.grid(row=1, column=1, padx=5, sticky="w")
 
-# Output directory selection
-output_dir_label = tk.Label(root, text="Select Output Directory:")
+output_dir_label = tk.Label(single_frame, text="Select Output Directory:")
 output_dir_label.grid(row=2, column=0, sticky="w")
-output_dir_entry = tk.Entry(root, width=50)
-output_dir_entry.grid(row=2, column=1, padx=5, pady=5)
-output_dir_button = tk.Button(root, text="Browse", command=select_output_dir)
+single_output_dir_entry = tk.Entry(single_frame, width=50)
+single_output_dir_entry.grid(row=2, column=1, padx=5, pady=5)
+output_dir_button = tk.Button(single_frame, text="Browse", command=select_single_output_dir)
 output_dir_button.grid(row=2, column=2, padx=5, pady=5)
 
-# Process data button
-process_button = tk.Button(root, text="Process Data", command=process_data)
-process_button.grid(row=4, column=1, pady=10)
+process_single_button = tk.Button(single_frame, text="Process Data", command=process_single_file, state=tk.DISABLED)
+process_single_button.grid(row=3, column=1, pady=10)
+
+# Folder Section
+folder_label = tk.Label(folder_frame, text="Select Folder:")
+folder_label.grid(row=0, column=0, sticky="w")
+folder_entry = tk.Entry(folder_frame, width=50)
+folder_entry.grid(row=0, column=1, padx=5, pady=5)
+folder_button = tk.Button(folder_frame, text="Browse", command=select_folder)
+folder_button.grid(row=0, column=2, padx=5, pady=5)
+
+mass_label_folder = tk.Label(folder_frame, text="Mass of Active Material (g):")
+mass_label_folder.grid(row=1, column=0, sticky="w")
+entry_mass_folder = tk.Entry(folder_frame, width=20)
+entry_mass_folder.grid(row=1, column=1, padx=5, sticky="w")
+
+output_dir_label_folder = tk.Label(folder_frame, text="Select Output Directory:")
+output_dir_label_folder.grid(row=2, column=0, sticky="w")
+folder_output_dir_entry = tk.Entry(folder_frame, width=50)
+folder_output_dir_entry.grid(row=2, column=1, padx=5, pady=5)
+output_dir_button_folder = tk.Button(folder_frame, text="Browse", command=select_folder_output_dir)
+output_dir_button_folder.grid(row=2, column=2, padx=5, pady=5)
+
+process_folder_button = tk.Button(folder_frame, text="Process Data", command=process_folder, state=tk.DISABLED)
+process_folder_button.grid(row=3, column=1, pady=10)
+
+# Close Button
+close_button = tk.Button(root, text="Close", command=on_closing)
+close_button.grid(row=3, column=0, pady=10)
 
 # Copyright Label
 copyright_label = tk.Label(root, text="Copyright Anurag Mohanty - 2024", fg="grey")
-copyright_label.grid(row=5, column=0,  sticky="w", padx=5)
+copyright_label.grid(row=5, column=0, sticky="w", padx=5)
 
+root.protocol("WM_DELETE_WINDOW", on_closing)
 root.mainloop()
